pyproject.toml
[project]
name = "upstage-network-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.2",
    "httpx==0.27.1", # 0.28.1 버전 호환 이슈
    "openai==1.52.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sse-starlette>=3.0.3",
    "chromadb>=0.5.0",
    "uvicorn[standard]>=0.38.0",
]
